### 2022-02-15 #4168 @whedon
[PRE REVIEW]: mxnorm: An R Package to Normalize Multiplexed Imaging Data
**Submitting author:** @ColemanRHarris (<a href="http://orcid.org/0000-0002-6325-0694">Coleman Harris</a>)
**Repository:** <a href="https://github.com/ColemanRHarris/mxnorm" target ="_blank">https://github.com/ColemanRHarris/mxnorm</a>
**Version:** v1.0.0
