### 2022-02-15 #4168 @danielskatz
👋 @ColemanRHarris - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). In addition, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references from branch joss-paper` to check again, and the command `@whedon generate pdf from branch joss-paper` after making changes to the .md file or when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
