+++ R/run_otsu_discordance.R
@@ -22,7 +22,8 @@
 #'   c("metadata1_vals"))
 #' mx_data = mx_normalize(mx_data, transform="log10",method="None")
 #' ## using `threshold_override` here in case users haven't installed `scikit-image`
-#' mx_data = run_otsu_discordance(mx_data, table="normalized", threshold_override = function(thold_data){quantile(thold_data, 0.5)})
+#' mx_data = run_otsu_discordance(mx_data, table="normalized",
+#' threshold_override = function(thold_data){quantile(thold_data, 0.5)})
 run_otsu_discordance <- function(mx_data,
                           table,
                           threshold_override = NULL,
