### 2022-02-18 #4180 @whedon
[REVIEW]: mxnorm: An R Package to Normalize Multiplexed Imaging Data
**Submitting author:** @ColemanRHarris (<a href="http://orcid.org/0000-0002-6325-0694">Coleman Harris</a>)
**Repository:** https://github.com/ColemanRHarris/mxnorm
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.2-joss
