+++ R/get_discordance_thold.R
@@ -5,14 +5,14 @@
 #' @return function to use as thresholding algorithm
 get_discordance_thold <- function(threshold_override,
                                ...){
-    skf <- reticulate::import("skimage.filters")
-
     ## return Otsu if null
     if(is.null(threshold_override)){
+        skf <- reticulate::import("skimage.filters")
         return(skf["threshold_otsu"])
     }
     ## return other skf module if character
     if(class(threshold_override) == "character"){
+        skf <- reticulate::import("skimage.filters")
         return(skf[paste0("threshold_",threshold_override)])
     }
     ## validate function if user-defined
