+++ R/utils-pipe.R
@@ -8,4 +8,5 @@
 #' @param lhs A value or the magrittr placeholder.
 #' @param rhs A function call using the magrittr semantics.
 #' @return The result of calling `rhs(lhs)`.
+#' @noRd
 NULL
