### 2022-02-22 4c4a9cc
Updated vignette for CRAN
