+++ doc/mxnorm-vignette.Rmd
@@ -1,0 +1,544 @@
+---
+title: The mxnorm package for removing slide-to-slide variation in multiplexed imaging
+  data
+author: |
+  | Coleman Harris
+  | Department of Biostatistics, Vanderbilt University Medical Center, Nashville, TN, USA
+output:
+  rmarkdown::html_vignette:
+    toc: yes
+    toc_depth: 3
+    fig_width: 7.3
+bibliography: mxnorm_vcitations.bib
+link-citations: yes
+vignette: >
+  %\VignetteIndexEntry{mxnorm-vignette1}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  message=FALSE,
+  warning=FALSE
+)
+```
+
+This vignette describes the `mxnorm` R package, and provides example analyses & detailed methodological background for using normalization methods in multiplexed imaging data. The package largely provides two key services: (1) a collection of normalization methods and analysis metrics to implement and compare normalization in multiplexed imaging data, and (2) a foundation for storing multiplexed imaging data in R using S3. Background for these methods and the multiplexed imaging field can be found in the author's previously published work [@harris2022quantifying].
+
+## Basic example
+
+This vignette using the following packages:
+
+```{r setup, warning=FALSE,message=FALSE}
+library(mxnorm)
+library(dplyr)
+library(reticulate)
+```
+
+We also use the Python module `scikit-image.filters` later on in this vignette -- the `mxnorm` package has been configured to try and avoid any conflicts with installing this module.
+
+```{r}
+if(reticulate::py_module_available("skimage")){
+  knitr::knit_engines$set(python = reticulate::eng_python)
+  skimage_available = TRUE
+} else{
+  ## set boolean for CRAN builds
+  skimage_available = FALSE
+  
+  ## install if running locally
+  #py_install("scikit-image")
+}
+```
+
+If you are prompted to install `Miniconda`, please do so. Please also see the `reticulate` [package documentation](https://rstudio.github.io/reticulate/articles/python_packages.html) for more information if you run into issues.
+
+### Loading in data
+In general, we expect multiplexed imaging data in a `data.frame` format that includes columns for slide & image identifiers, separate columns for marker intensity values, and some set of metadata columns like tissue identifiers, phenotypic traits, medical conditions, etc. Alongside the `mxnorm` package itself, we introduce the `mx_sample` dataset that demonstrates the expected structure of multiplexed imaging data, and provides simulated marker intensity values that demonstrate strong slide effects. This structure is as follows:
+
+```{r}
+head(mx_sample, 3)
+```
+This dataset consists of 3 markers across 4 slides (with 750 "cells" on each slide) and 1 metadata column, and was specifically created to demonstrate the effect of normalization methods in multiplexed imaging data. To ensure a streamlined framework for the analysis of this type of data, we have created an S3 object `mx_dataset` to store the data and continue building upon as we normalize the data and analyze our results. 
+
+### Creating the S3 object with `mx_dataset()`
+Let's load the `mx_sample` dataset into the S3 object we'll use for our analyses, the `mx_dataset` object. Here we specify:
+
+- `data`: the input dataset in a `data.frame` format
+- `slide_id` and `image_id`: the identifiers of interest in our dataset
+- `marker_cols`: the set of marker columns in our input data that we want to include
+- `metadata_cols`: metadata columns in our input data that we want to include
+
+```{r example}
+mx_data = mx_dataset(data=mx_sample,
+                        slide_id="slide_id",
+                        image_id="image_id",
[…]
