+++ paper.md
@@ -27,7 +27,7 @@ link-citations: true
 
 Multiplexed imaging is an emerging single-cell assay that can be used to understand and analyze complex processes in tissue-based cancers, autoimmune disorders, and more. These imaging technologies, which include co-detection by indexing (CODEX), multiplexed ion beam imaging (MIBI), and multiplexed immunofluorescence imaging (MxIF), provide detailed information about spatial interactions between cells [@goltsev2018deep; @angelo2014multiplexed; @gerdes2013highly]. Multiplexed imaging experiments generate data across hundreds of slides and images, often resulting in terabytes of complex data to analyze through imaging analysis pipelines. Methods are rapidly developing to improve particular parts of the pipeline, including software packages in R and Python like `spatialTime`, `imcRtools`, `MCMICRO`, and `Squidpy` [@creed2021spatialtime; @windhager2021end; @schapiro2021mcmicro; @palla2021squidpy]. An important, but understudied component of this pipeline is the analysis of technical variation within this complex data source --  intensity normalization is one way to remove this technical variability. The combination of disparate pre-processing pipelines, imaging variables, optical effects, and within-slide dependencies create batch and slide effects that can be reduced via normalization methods. Current state-of-the-art methods vary heavily across research labs and image acquisition platforms, without one singular method that is uniformly robust -- optimal statistical methods seek to improve similarity across images and slides by removing this technical variability while maintaining the underlying biological signal in the data.
 
-`mxnorm` is open-source software built with R and S3 methods that implements, evaluates, and visualizes normalization techniques for multiplexed imaging data. Extending methodology described in @harris2022quantifying, we intend to set a foundation for the evaluation of multiplexed imaging normalization methods in R. This easily allows users to extend normalization methods into the field, and provides a robust evaluation framework to measure both technical variability and the efficacy of various normalization methods. One key component of the R package is the ability to supply user-defined normalization methods and thresholding algorithms to assess normalization in multiplexed imaging data. Core features, usage details, and extensive tutorials are available in the [package documentation and vignette](https://github.com/ColemanRHarris/mxnorm).
+`mxnorm` is open-source software built with R and S3 methods that implements, evaluates, and visualizes normalization techniques for multiplexed imaging data. Extending methodology described in @harris2022quantifying, we intend to set a foundation for the evaluation of multiplexed imaging normalization methods in R. This easily allows users to extend normalization methods into the field, and provides a robust evaluation framework to measure both technical variability and the efficacy of various normalization methods. One key component of the R package is the ability to supply user-defined normalization methods and thresholding algorithms to assess normalization in multiplexed imaging data. Core features, usage details, and extensive tutorials are available in the [package documentation and vignette on CRAN](https://cran.r-project.org/package=mxnorm) and [the software repository](https://github.com/ColemanRHarris/mxnorm).
 
 # Statement of need
 
@@ -41,7 +41,7 @@ We recently proposed and evaluated several normalization methods for multiplexed
 
 ![Figure 1: Basic structure of the `mxnorm` package and associated functions](mxnorm_structure.png){width=100%}
 
[…]
