### 2022-02-25 9f76d74
finished fixes for CRAN
