+++ README.md
@@ -13,7 +13,9 @@
 
 A package designed to handle multiplexed imaging data in R, implementing
 normalization methods and quality metrics detailed in our paper
-[here](https://doi.org/10.1101/2021.07.16.452359).
+[here](https://doi.org/10.1101/2021.07.16.452359). Further information
+about the package, usage, the vignettes, and more can be found on
+[CRAN](https://cran.r-project.org/package=mxnorm).
 
 # Installation
 
@@ -56,7 +58,7 @@ mx_dataset = mx_dataset(data=mx_sample,
                         metadata_cols=c("metadata1_vals"))
 ```
 
-WE can use the built-in `summary()` function to observe `mx_dataset`
+We can use the built-in `summary()` function to observe `mx_dataset`
 object:
 
 ``` r
@@ -75,20 +77,7 @@ mx_norm = mx_normalize(mx_data = mx_dataset,
                        method="None")
 ```
 
-The `mx_dataset` object has normalized data in the following form:
-
-``` r
-head(mx_norm$norm_data)
-#>   slide_id image_id marker1_vals marker2_vals marker3_vals metadata1_vals
-#> 1   slide1   image1    0.1991146    0.1553918   0.18315112            yes
-#> 2   slide1   image1    0.1082691    0.2294791   0.21339780             no
-#> 3   slide1   image1    0.1328231    0.1389272   0.11549488            yes
-#> 4   slide1   image1    0.1082691    0.1865590   0.23245192            yes
-#> 5   slide1   image1    0.1328231    0.2156384   0.13923749            yes
-#> 6   slide1   image1    0.1082691    0.1553918   0.07725425            yes
-```
-
-And we can use `summary()` to capture the following attributes for the
+And we again use `summary()` to capture the following attributes for the
 `mx_dataset` object:
 
 ``` r
@@ -119,48 +108,25 @@ mx_otsu = run_otsu_discordance(mx_norm,
                         plot_out = FALSE)
 ```
 
-This adds an Otsu misclassification table to the `mx_dataset` object in
-the following form:
+We can also begin to visualize these results using some of `mxnorm`’s
+plotting features built using `ggplot2`.
+
+First, we can visualize the densities of the marker values as follows:
 
 ``` r
-head(mx_otsu$otsu_data)
-#>   slide_id       marker table slide_threshold marker_threshold
-#> 1   slide1 marker1_vals   raw        12.01758         54.89844
-#> 2   slide2 marker1_vals   raw        20.01367         54.89844
-#> 3   slide3 marker1_vals   raw        87.05664         54.89844
-#> 4   slide4 marker1_vals   raw        44.00391         54.89844
-#> 5   slide1 marker2_vals   raw        19.00977         52.90039
-#> 6   slide2 marker2_vals   raw        19.99219         52.90039
-#>   discordance_score
-#> 1         0.4506667
-#> 2         0.4306667
-#> 3         0.2573333
-#> 4         0.3386667
-#> 5         0.5333333
-#> 6         0.5320000
+plot_mx_density(mx_otsu)
 ```
 
-And we can use `summary()` to capture the following attributes for the
-`mx_dataset` object:
+<img src="man/figures/README-mx_dens-1.png" width="100%" />
+
+We can also visualize the results of the Otsu misclassification analysis
+stratified by slide and marker:
 
 ``` r
-summary(mx_otsu)
-#> Call:
-#> `mx_dataset` object with 4 slide(s), 3 marker column(s), and 1 metadata column(s)
-#> 
-#> Normalization:
-#> Data normalized with transformation=`log10_mean_divide` and method=`None`
-#> 
-#> Anderson-Darling tests:
-#>       table mean_test_statistic mean_std_test_statistic mean_p_value
-#>  normalized              34.565                  24.111            0
[…]
