### 2022-02-25 51647b8
fix CRAN note on classes
