### 2022-03-12 ColemanRHarris/mxnorm#4 @tijeco
Installation notes
I've tried installing mxnorm on a fresh environment. Here is what my setup was so that you have a better idea.

I'm using a freshly installed Linux container installed on a pixelbook (which is a docker container of a fairly bare bones Debian, I can provide the specific version if necessary). I installed miniconda, and made a separate environment for R using the following command (also installed mamba, because mamba is amazing):

```
mamba create -n R -c conda-forge r-base r-essentials
```

Inside the R environment, I followed the instructions to install mxnorm ` install.packages("mxnorm")`

On my end, it gave an error that `lme4` and `nloptr` needed to be installed. I used the following command to get those installed through conda:

```
mamba install -c conda-forge r-lme4 r-nloptr
```

Now, my macbook already has a well established R environment setup, so I doubt I would have any problems installing it there, I just wanted to see how it would work on a fresh environment.

I just wanted to bring those dependencies to your attention. I think it should be a pretty easy thing to either include in the README or on CRAN setup side of things.

I also wanted to mention, in case you haven't considered it already, to at some point make mxnorm available through conda. It seems like it wouldn't be too difficult, and most of the dependencies should already be available through there.
