### 2022-03-12 #4180 @tijeco
@osorensen @ColemanRHarris I'm going through this now. I've installed the software in a fresh environment and have a few notes / suggestions for the author to include in the documentation. On my end, it wouldn't install unless lme4 and nloptr were also installed. I've raised this as an issue on the mxnorm repo (https://github.com/ColemanRHarris/mxnorm/issues/4)
