### 2022-03-12 ColemanRHarris/mxnorm#5 @tijeco
automated testing
It seems that you have a set of test scripts, but I don't see where them being implemented in some sort of continuous integration framework. You can set up a reasonably simple github actions pipeline to run those tests each time you make a commit. It's a good habit to get into, because it can provide immediate input about potential bugs that you have introduced that may cause a simple test to fail.
