### 2022-03-12 #4180 @tijeco
@osorensen @ColemanRHarris I don't see automated tests. I see the test scripts though! I've raised an issue here (https://github.com/ColemanRHarris/mxnorm/issues/5)
