### 2022-03-12 ColemanRHarris/mxnorm#6 @tijeco
Interesting issue when running in conda environment
As mentione in #4, I've installed mxnorm on a fresh computer in a conda environment.

I was going through the examples in the README and ran into this when I executed the follwowing:

```R
mx_otsu = run_otsu_discordance(mx_norm,
                        table="both",
                        threshold_override = NULL,
                        plot_out = FALSE)
```

The output was:

```
No non-system installation of Python could be found.
Would you like to download and install Miniconda?
Miniconda is an open source environment management system for Python.
See https://docs.conda.io/en/latest/miniconda.html for more details.

Would you like to install Miniconda? [Y/n]
```

For context, outside of conda, I don't have python installed (there is no /usr/bin/python).

In the conda environment that I am using and have mxnorm installed in, python is installed, and should be accessible through the $PATH.

Also, I already have miniconda installed, so I certainly wouldn't want to proceed with installing it again.

I think some of this should be mentioned in the documentation, as far as dependencies go.

Also, is there a way to determine that miniconda is installed, or that mxnorm is already installed in a conda environment?
