### 2022-03-12 #4180 @tijeco
@osorensen @ColemanRHarris I ran into an interesting issue when going through the examples in the README, I've raised a separate issue (https://github.com/ColemanRHarris/mxnorm/issues/6).
