### 2022-03-14 #4180 @osorensen
@askerdb, could you try running `@editorialbot generate my checklist`? It should generate a new checklist which you can use instead.
