### 2022-03-14 ColemanRHarris/mxnorm#7 @askerdb
References to underlying software
The paper does not have a methods section or any other place with references to the underlying packages, lme4, umap etc. that all deserve to be cited
