### 2022-03-14 #4180 @askerdb
I added an issue about references to the underlying software https://github.com/ColemanRHarris/mxnorm/issues/7
and am waiting for clarification on the automated tests and contributors file the other reviewer raised. The software installed (from cran) and ran perfectly for me.
