### 2022-03-14 1bdc26e
fixed error with tibble input
