### 2022-03-16 ColemanRHarris/mxnorm#4 @ColemanRHarris
After doing some digging, it looks like both `lme4` (and `nloptr` since it's one of the dependencies) are not included in the `r-essentials` bundle from conda. `lme4` is listed as an "Import" in the `DESCRIPTION` file, so I would expect that R would either prompt you to install those packages or go ahead and install them by default when installing `mxnorm`.

Can you share more info about the error you encountered? Was it the R environment flagging that you needed to install those packages, and instead of installing them within the R environment you installed them via `conda`?
