+++ README.md
@@ -9,6 +9,7 @@
 [![](http://cranlogs.r-pkg.org/badges/grand-total/mxnorm?color=green)](https://cran.r-project.org/package=mxnorm)
 [![status](https://joss.theoj.org/papers/c18a1df622016c6ea8ed9e67dda73d07/status.svg)](https://joss.theoj.org/papers/c18a1df622016c6ea8ed9e67dda73d07)
 
+[![R-CMD-check](https://github.com/ColemanRHarris/mxnorm/workflows/R-CMD-check/badge.svg)](https://github.com/ColemanRHarris/mxnorm/actions)
 <!-- badges: end -->
 
 A package designed to handle multiplexed imaging data in R, implementing
@@ -205,8 +206,8 @@ summary(mx_var)
 #> 
 #> Clustering consistency (UMAP):
 #>       table adj_rand_index cohens_kappa
-#>  normalized          0.052       -0.102
-#>         raw          0.574        0.206
+#>  normalized          0.052        0.090
+#>         raw          0.633        0.522
 #> 
 #> Variance proportions (slide-level):
 #>       table  mean    sd
