### 2022-03-16 ColemanRHarris/mxnorm#5 @ColemanRHarris
Automated testing for both the `R CMD check` and code coverage has been added via the Github Actions pipeline in the commit here: 62e9b45756dd3bce559ce8d73307e0f2242cd6c9
