### 2022-03-16 #4180 @ColemanRHarris
> @osorensen @ColemanRHarris I'm not sure if this needs its own separate issue, but I don't see a contributing section.

A "Community Guidelines" section has been added to the package README here: https://github.com/ColemanRHarris/mxnorm#community-guidelines
