### 2022-03-16 ColemanRHarris/mxnorm#4 @ColemanRHarris
So I was able to setup a similar Docker container and re-create this issue, which turned out to be a missing dependency for `CMake` installed on the machine. I've added this to the [Dependencies](https://github.com/ColemanRHarris/mxnorm#dependencies) section of the README in case users hit the same error.
