### 2022-03-16 ColemanRHarris/mxnorm#6 @ColemanRHarris
As discussed in https://github.com/ColemanRHarris/mxnorm/issues/4, I was able to setup a similar Docker container and re-create this issue. Based on some digging in the `conda` repository, this was a known issue (https://github.com/rstudio/reticulate/issues/607) and was seemingly addressed. The steps I took to work around it were:
- Did not install Miniconda via the R console.
- Used `reticulate::py_config()` to setup my desired Python environment.
- Ensured `scikit-image` was installed in the same location as the Python environment.

These steps allowed me to run the command you listed above without error, and have been recreated in the [Dependencies](https://github.com/ColemanRHarris/mxnorm#dependencies) section of the README.
