### 2022-03-17 05e7a5b
added methods to paper
