+++ paper.md
@@ -41,7 +41,9 @@ We recently proposed and evaluated several normalization methods for multiplexed
 
 ![Figure 1: Basic structure of the `mxnorm` package and associated functions](mxnorm_structure.png){width=100%}
 
-As shown in **Figure 1**, there are three main types of functions implemented in the `mxnorm` package -- infrastructure, analysis, and visualization. The first infrastructure function, `mx_dataset()`, specifies and creates the S3 object used throughout the analysis, while the `mx_normalize()` function provides a routine to normalize the multiplexed imaging data, which specifically allows for normalization algorithms defined by the user. Each of the three analysis functions provides methods to run specific analyses that test for slide-to-slide variation and preservation of biological signal for the normalized and unnormalized data, while the four visualization functions provide methods to generate `ggplot2` plots to assess the results. We also extend the `summary()` generic function to the `mx_dataset` S3 object to provide further statistics and summaries. All of the statistical methodology behind these normalization and analysis methods are detailed further in our package [vignette](https://github.com/ColemanRHarris/mxnorm) and in the methods paper [@harris2022quantifying].
+As shown in **Figure 1**, there are three main types of functions implemented in the `mxnorm` package -- infrastructure, analysis, and visualization. The first infrastructure function, `mx_dataset()`, specifies and creates the S3 object used throughout the analysis, while the `mx_normalize()` function provides a routine to normalize the multiplexed imaging data, which specifically allows for normalization algorithms defined by the user. Each of the three analysis functions provides methods to run specific analyses that test for slide-to-slide variation and preservation of biological signal for the normalized and unnormalized data, while the four visualization functions provide methods to generate `ggplot2` plots to assess the results. We also extend the `summary()` generic function to the `mx_dataset` S3 object to provide further statistics and summaries. 
+
+The statistical methodology underlying the methods we implemented here in `mxnorm` builds upon existing work in both R and Python. Normalization algorithms available in `mx_normalize()` leverage methodology derived in the ComBat paper, the `fda` package, and the `tidyverse` framework [@johnson2007adjusting, @fda_pack, @wickham2019welcome]. The threshold discordance methods available in `run_otsu_discordance()` leverage methodology from Otsu's original paper and the `scikit-image` implementation of Otsu thresholding in Python [@otsu1979threshold, @scikit-image]. Our implementation of the UMAP algorithm in `run_reduce_umap()` leverages both the UMAP paper and the `uwot` implementation of the UMAP algorithm in R [@mcinnes2018umap, @uwot_pack]. The random effects modeling options available in `run_var_proportions()` leverage the `lme4` R package [@bates2015walker]. Even more information for the statistical methodology behind these normalization and analysis methods are detailed further in our package [vignette](https://cran.r-project.org/web/packages/mxnorm/vignettes/mxnorm-vignette.html) and in the methods paper [@harris2022quantifying].
 
 # A minimal example
 
