+++ mxnorm_pcitations.bib
@@ -3,8 +3,7 @@
   author={Harris, Coleman R and McKinley, Eliot T and Roland, Joseph T and Liu, Qi and Shrubsole, Martha J and Lau, Ken S and Coffey, Robert J and Wrobel, Julia and Vandekar, Simon N},
   journal={Bioinformatics (Oxford, England)},
   pages={btab877},
-  year={2022},
-	doi={10.1093/bioinformatics/btab877}
+  year={2022}
 }
 
 @article{ptacek2020multiplexed,
@@ -15,8 +14,7 @@
   number={8},
   pages={1111--1123},
   year={2020},
-  publisher={Nature Publishing Group},
-	doi={10.1038/s41374-020-0417-4}
+  publisher={Nature Publishing Group}
 }
 @article{gerdes2013highly,
   title={Highly multiplexed single-cell analysis of formalin-fixed, paraffin-embedded cancer tissue},
@@ -26,8 +24,7 @@
   number={29},
   pages={11982--11987},
   year={2013},
-  publisher={National Acad Sciences},
-	doi={10.1073/pnas.1300136110}
+  publisher={National Acad Sciences}
 }
 @article{chen2021differential,
   title={Differential pre-malignant programs and microenvironment chart distinct paths to malignancy in human colorectal polyps},
@@ -37,8 +34,7 @@
   number={26},
   pages={6262--6280},
   year={2021},
-  publisher={Elsevier},
-	doi={10.1016/j.cell.2021.11.031}
+  publisher={Elsevier}
 }
 
 @article{mckinley2022miriam,
@@ -46,8 +42,7 @@
   author={McKinley, Eliot T and Shao, Justin and Ellis, Samuel T and Heiser, Cody N and Roland, Joseph T and Macedonia, Mary C and Vega, Paige N and Shin, Susie and Coffey, Robert J and Lau, Ken S},
   journal={Cytometry Part A},
   year={2022},
-  publisher={Wiley Online Library},
-	doi={10.1002/cyto.a.24541}
+  publisher={Wiley Online Library}
 }
 @article{schapiro2021mcmicro,
   title={MCMICRO: A scalable, modular image-processing pipeline for multiplexed tissue imaging},
@@ -55,8 +50,7 @@
   journal={Nature methods},
   pages={1--5},
   year={2021},
-  publisher={Nature Publishing Group},
-	doi={10.1101/2021.03.15.435473}
+  publisher={Nature Publishing Group}
 }
 @article{graf2022flino,
   title={FLINO: a new method for immunofluorescence bioimage normalization},
@@ -66,8 +60,7 @@
   number={2},
   pages={520--526},
   year={2022},
-  publisher={Oxford University Press},
-	doi={10.1093/bioinformatics/btab686}
+  publisher={Oxford University Press}
 }
 @article{chang2020restore,
   title={RESTORE: Robust intEnSiTy nORmalization mEthod for multiplexed imaging},
@@ -77,8 +70,7 @@
   number={1},
   pages={1--9},
   year={2020},
-  publisher={Nature Publishing Group},
-	doi={10.1038/s42003-020-0828-1}
+  publisher={Nature Publishing Group}
 }
 
 @article{love2014moderated,
@@ -89,8 +81,7 @@
   number={12},
   pages={1--21},
   year={2014},
-  publisher={BioMed Central},
-	doi={10.1186/s13059-014-0550-8}
+  publisher={BioMed Central}
 }
 
 @incollection{smyth2005limma,
@@ -99,8 +90,29 @@
   booktitle={Bioinformatics and computational biology solutions using R and Bioconductor},
   pages={397--420},
   year={2005},
-  publisher={Springer},
-	doi={10.1007/0-387-29362-0_23}
+  publisher={Springer}
+}
+
[…]
