### 2022-03-17 ColemanRHarris/mxnorm#7 @ColemanRHarris
Methods and additional references to underlying packages have been added to the "Functionality" section of the paper in the following commit: 05e7a5b760414a611ed1b8d3e2b23e4ea61f53fc.
