### 2022-03-17 #4180 @ColemanRHarris
I believe that I’ve addressed all of the open issues:
- Added information about dependencies and installation instructions in the [Dependencies](https://github.com/ColemanRHarris/mxnorm#dependencies) section of the README.
- Added automated testing via Github Actions.
- Added a [Community Guidelines](https://github.com/ColemanRHarris/mxnorm#community-guidelines) section to the README.
- Added details about methods used in the `mxnorm` package to the Functionality section of the paper text.

Thank you @askerdb and @tijeco for your excellent feedback. What else do you need from me @osorensen?
