### 2022-03-17 #4180 @danielskatz
(new version of commands for new and improved bot)
