### 2022-03-18 2b808f6
remove checking method_override on mx_sample
