### 2022-03-26 #4180 @osorensen
@ColemanRHarris, before I can recommend acceptance, please fix the following issues:

- Add the missing DOIs stated in the post above.
- The term "Figure 1" is duplicated in the caption to Figure 1. Please fix.
- Be consistent, and use full journal names for all references. For example, on line 111, use Journal of Statistical Software rather than J Stat Softw.
- Make sure R is properly capitaliizied in the references, e.g., on line 176. You can enforce this by adding an extra set of curly braces in the `.bib` source file. For this particular example, that would mean writing `booktitle={Bioinformatics and computational biology solutions using {R} and {Bioconductor}}.`

Otherwise, congratulations with a very good and useful R package!
