+++ mxnorm_pcitations.bib
@@ -1,9 +1,10 @@
 @article{harris2022quantifying,
   title={Quantifying and correcting slide-to-slide variation in multiplexed immunofluorescence images},
   author={Harris, Coleman R and McKinley, Eliot T and Roland, Joseph T and Liu, Qi and Shrubsole, Martha J and Lau, Ken S and Coffey, Robert J and Wrobel, Julia and Vandekar, Simon N},
-  journal={Bioinformatics (Oxford, England)},
+  journal={Bioinformatics},
   pages={btab877},
-  year={2022}
+  year={2022},
+	doi={10.1093/bioinformatics/btab877}
 }
 
 @article{ptacek2020multiplexed,
@@ -14,7 +15,8 @@
   number={8},
   pages={1111--1123},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+	doi={10.1038/s41374-020-0417-4}
 }
 @article{gerdes2013highly,
   title={Highly multiplexed single-cell analysis of formalin-fixed, paraffin-embedded cancer tissue},
@@ -24,7 +26,8 @@
   number={29},
   pages={11982--11987},
   year={2013},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+	doi={10.1073/pnas.1300136110}
 }
 @article{chen2021differential,
   title={Differential pre-malignant programs and microenvironment chart distinct paths to malignancy in human colorectal polyps},
@@ -34,7 +37,8 @@
   number={26},
   pages={6262--6280},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+	doi={10.1016/j.cell.2021.11.031}
 }
 
 @article{mckinley2022miriam,
@@ -42,15 +46,17 @@
   author={McKinley, Eliot T and Shao, Justin and Ellis, Samuel T and Heiser, Cody N and Roland, Joseph T and Macedonia, Mary C and Vega, Paige N and Shin, Susie and Coffey, Robert J and Lau, Ken S},
   journal={Cytometry Part A},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+	doi={10.1002/cyto.a.24541}
 }
 @article{schapiro2021mcmicro,
   title={MCMICRO: A scalable, modular image-processing pipeline for multiplexed tissue imaging},
   author={Schapiro, Denis and Sokolov, Artem and Yapp, Clarence and Chen, Yu-An and Muhlich, Jeremy L and Hess, Joshua and Creason, Allison L and Nirmal, Ajit J and Baker, Gregory J and Nariya, Maulik K and others},
-  journal={Nature methods},
+  journal={Nature Methods},
   pages={1--5},
   year={2021},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+	doi={10.1101/2021.03.15.435473}
 }
 @article{graf2022flino,
   title={FLINO: a new method for immunofluorescence bioimage normalization},
@@ -60,92 +66,77 @@
   number={2},
   pages={520--526},
   year={2022},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+	doi={10.1093/bioinformatics/btab686}
 }
 @article{chang2020restore,
   title={RESTORE: Robust intEnSiTy nORmalization mEthod for multiplexed imaging},
   author={Chang, Young Hwan and Chin, Koei and Thibault, Guillaume and Eng, Jennifer and Burlingame, Erik and Gray, Joe W},
-  journal={Communications biology},
+  journal={Communications Biology},
   volume={3},
   number={1},
   pages={1--9},
   year={2020},
[…]
