+++ mxnorm_pcitations.bib
@@ -253,7 +253,8 @@
   number={1},
   pages={62--66},
   year={1979},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tsmc.1979.4310076}
 }
 
 @article{scikit-image,
