+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: mxnorm
 Title: Apply Normalization Methods to Multiplexed Images
-Version: 1.0.1
+Version: 1.0.2
 Authors@R: 
     person(given = "Coleman",
            family = "Harris",
