### 2022-03-28 #4180 @ColemanRHarris
Hi @osorensen! I have completed the above tasks:

- The version tag of the archived version is `v1.0.2-joss`.
- The DOI of the archived version is `10.5281/zenodo.6390746`.
