### 2022-03-30 #4180 @kyleniemeyer
Hi @ColemanRHarris, I'm doing some final checks before publishing.

Could you add the appropriate `arXiv:YYMM.NNNNv# [category]` to the McInnes reference, and any other arXiv references I may have missed? You can do this in the `.bib` file with `{\tt arXiv:0706.1234 [math.FA]}` (for example)
