+++ mxnorm_pcitations.bib
@@ -287,7 +287,7 @@ for Dimensionality Reduction},
   @article{mcinnes2018umap,
   title={Umap: Uniform manifold approximation and projection for dimension reduction},
   author={McInnes, Leland and Healy, John and Melville, James},
-  journal={arXiv},
+  journal={arXiv preprint arXiv:1802.03426 [stat.ML]},
   year={2018},
 	doi={10.48550/arXiv.1802.03426}
 }
@@ -300,5 +300,5 @@ for Dimensionality Reduction},
   number={1},
   pages={1--48},
   year={2015},
-	doi={10.48550/arXiv.1406.5823}
+	doi={10.18637/jss.v067.i01}
 }
